Move sensor climate functions to own climate.py and all coffee things to coffee.py
